refactor(kv-cache): move indexer state spec to core
Place the reusable Ascend indexer K-pool state specification beside the other Ascend KV cache specs. Keep GLM-specific model metadata on the model-owned cache layer and remove the core-to-model registration dependency.
